\name{ballsdMN-package}
\alias{ballsdMN-package}
\alias{ballsdMN}
\docType{package}
\title{
 Optimization by Bayesian adaptive locally linear stochastic descent
}
\description{
  Bayesian Adaptive Locally Linear Stochastic Descent (BALLSD) has been designed to replicate the essential aspects of manual parameter fitting in an automated way. Specifically, BALLSD uses simple Bayesian principles to form probabilistic assumptions about (a) which parameters have the greatest effect on the objective function, and (b) optimal step sizes for each parameter.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
My Nguyen (Jasmine).

Maintainer: Your Name <nguyenm21@vcu.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
